+++ packages/plugin-dev-essentials/src/bin/postpack.sh
@@ -2,6 +2,8 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 # DEV_ESSENTIALS_DIR="$(realpath "${DIR}/../..")"
 # CURRENT_DIR="$(pwd)"
 
+echo "RUNNING POSTPACK IN '${DIR}'"
+
 # commands resolution
 jq '.oclif.commands="./src/commands"' package.json | sponge package.json