+++ packages/pkg-error/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @business-as-code/error
 
+## 0.0.0-latest-20230812071138
+
+### Patch Changes
+
+- this is a snapshot release
+
+## 0.0.0-latest-20230812070809
+
+### Patch Changes
+
+- this is a snapshot release
+
+## 0.0.0-latest-20230812065455
+
+### Patch Changes
+
+- this is a snapshot release
+
+## 0.0.0-latest-20230812065321
+
+### Patch Changes
+
+- this is a snapshot release
+
+## 0.0.0-latest-20230812064733
+
+### Patch Changes
+
+- this is a snapshot release
+
+## 0.0.0-latest-20230812064324
+
+### Patch Changes
+
+- this is a snapshot release
+
 ## 0.0.0-latest-20230811111950
 
 ### Patch Changes